move storybook configs to `configs/storybook`
