+++ apps/storybooks/package.json
@@ -7,6 +7,7 @@
 		"dev": "wireit"
 	},
 	"devDependencies": {
+		"storybook": "8.0.5",
 		"wireit": "0.14.4"
 	},
 	"wireit": {
