+++ configs/storybook/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "@configs/storybook",
+	"version": "0.1.0",
+	"private": true,
+	"type": "module",
+	"dependencies": {
+		"@babel/core": "7.24.0",
+		"@babel/plugin-transform-typescript": "7.24.5",
+		"@babel/preset-env": "7.24.0",
+		"@babel/preset-react": "7.24.1",
+		"@babel/preset-typescript": "7.24.1",
+		"@emotion/babel-plugin": "11.11.0",
+		"@emotion/react": "11.11.1",
+		"@guardian/source-foundations": "workspace:*",
+		"@storybook/addon-a11y": "8.0.5",
+		"@storybook/addon-docs": "8.0.5",
+		"@storybook/addon-essentials": "8.0.5",
+		"@storybook/addon-links": "8.0.5",
+		"@storybook/addon-viewport": "8.0.5",
+		"@storybook/manager-api": "8.0.5",
+		"@storybook/react": "8.0.5",
+		"@storybook/react-webpack5": "8.0.5",
+		"@storybook/theming": "8.0.5",
+		"@types/babel__core": "7.20.5",
+		"@types/react": "18.2.11",
+		"babel-loader": "9.1.3",
+		"react": "18.2.0"
+	}
+}
