+++ configs/storybook/preview/FocusManagerDecorator.tsx
@@ -1,6 +1,6 @@
-import { FocusStyleManager } from '@guardian/source-foundations';
 import type { Decorator } from '@storybook/react';
 import { useEffect } from 'react';
+import { FocusStyleManager } from '@guardian/source-foundations';
 
 export const FocusManagerDecorator: Decorator = (storyFn) => {
 	useEffect(() => {
