+++ libs/@guardian/source-react-components-development-kitchen/.storybook/preview.ts
@@ -1,0 +1,1 @@
+export { decorators, parameters } from '../../../../configs/storybook/preview';
