+++ libs/@guardian/source-react-components/.storybook/preview.ts
@@ -1,0 +1,2 @@
+export { decorators } from '../../../../configs/storybook/preview';
+export { parameters } from '../../../../configs/storybook/preview';
