--- libs/@guardian/source-react-components/.storybook/preview/FocusManagerDecorator.tsx
@@ -1,11 +1,0 @@
-import { FocusStyleManager } from '@guardian/source-foundations';
-import type { Decorator } from '@storybook/react';
-import { useEffect } from 'react';
-
-export const FocusManagerDecorator: Decorator = (storyFn) => {
-	useEffect(() => {
-		FocusStyleManager.onlyShowFocusOnTabs();
-	});
-
-	return <>{storyFn()}</>;
-};
