--- libs/@guardian/source-react-components/.storybook/preview/ThemeProviderDecorator.tsx
@@ -1,13 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment -- storybook type contains `any`s */
-
-import { ThemeProvider } from '@emotion/react';
-import type { Decorator } from '@storybook/react';
-
-export const ThemeProviderDecorator: Decorator = (storyFn, context) => {
-	const theme = context.parameters.theme;
-	return theme ? (
-		<ThemeProvider theme={theme}>{storyFn()}</ThemeProvider>
-	) : (
-		<>{storyFn()}</>
-	);
-};
