--- libs/@guardian/source-react-components/.storybook/preview/backgrounds.ts
@@ -1,27 +1,0 @@
-import { palette } from '@guardian/source-foundations';
-
-export const backgrounds = {
-	default: 'background.primary',
-	values: [
-		{
-			name: 'background.primary',
-			value: palette.neutral[100],
-		},
-		{
-			name: 'background.secondary',
-			value: palette.neutral[97],
-		},
-		{
-			name: 'background.inverse',
-			value: palette.neutral[10],
-		},
-		{
-			name: 'brandBackground.primary',
-			value: palette.brand[400],
-		},
-		{
-			name: 'brandAltBackground.primary',
-			value: palette.brandAlt[400],
-		},
-	],
-};
