+++ libs/@guardian/source-react-components/package.json
@@ -29,6 +29,7 @@
 		"@babel/core": "7.24.0",
 		"@emotion/react": "11.11.1",
 		"@guardian/source-foundations": "16.0.0",
+		"@storybook/manager-api": "8.0.5",
 		"@svgr/babel-preset": "8.1.0",
 		"@svgr/core": "8.1.0",
 		"@svgr/plugin-jsx": "8.1.0",
@@ -44,6 +45,7 @@
 		"prettier": "3.2.2",
 		"react": "18.2.0",
 		"rollup": "4.17.2",
+		"storybook": "8.0.5",
 		"ts-jest": "29.1.1",
 		"ts-node": "10.9.2",
 		"tslib": "2.6.2",
@@ -94,6 +96,8 @@
 				"_deps"
 			],
 			"files": [
+				"../../../../configs/storybook/**",
+				"./.storybook/**",
 				"**/*.{ts,tsx,js,json,mdx,md}",
 				"!(dist|storybook-static)/**",
 				"!(jest*)"
