--- libs/@guardian/source/.storybook/preview.tsx
@@ -1,23 +1,0 @@
-import { backgrounds } from './preview/backgrounds';
-import { FocusManagerDecorator } from './preview/FocusManagerDecorator';
-import { ThemeProviderDecorator } from './preview/ThemeProviderDecorator';
-import { viewport } from './preview/viewport';
-
-export const parameters = {
-	viewport,
-	backgrounds,
-	actions: { argTypesRegex: '^on[A-Z].*' },
-	controls: {
-		matchers: {
-			color: /(background|color)$/i,
-			date: /Date$/,
-		},
-	},
-	options: {
-		storySort: {
-			method: 'alphabetical',
-		},
-	},
-};
-
-export const decorators = [FocusManagerDecorator, ThemeProviderDecorator];
