--- libs/@guardian/source/.storybook/theme.ts
@@ -1,9 +1,0 @@
-import { create } from '@storybook/theming';
-
-export const theme = create({
-	base: 'light',
-	brandTitle: '@guardian/source',
-	brandUrl: 'https://www.npmjs.com/package/@guardian/source',
-	brandImage:
-		'https://raw.githubusercontent.com/guardian/source/main/assets/logo.png',
-});
