+++ libs/@guardian/source/package.json
@@ -41,6 +41,7 @@
 		"@emotion/react": "11.11.1",
 		"@guardian/design-tokens": "workspace:*",
 		"@rollup/plugin-alias": "5.1.0",
+		"@storybook/manager-api": "8.0.5",
 		"@svgr/babel-preset": "8.1.0",
 		"@svgr/core": "8.1.0",
 		"@svgr/plugin-jsx": "8.1.0",
@@ -57,6 +58,7 @@
 		"prettier": "3.2.2",
 		"react": "18.2.0",
 		"rollup": "4.17.2",
+		"storybook": "8.0.5",
 		"ts-jest": "29.1.1",
 		"tslib": "2.6.2",
 		"tsx": "4.7.1",
@@ -104,6 +106,8 @@
 		"build-storybook": {
 			"command": "NODE_ENV=production storybook build --webpack-stats-json",
 			"files": [
+				"../../../../configs/storybook/**",
+				"./.storybook/**",
 				"**/*.{ts,tsx,js,json,mdx,md}",
 				"!storybook-static/**",
 				"!jest*"
