+++ package.json
@@ -9,35 +9,18 @@
 		"test": "./tools/scripts/use-make-instead"
 	},
 	"devDependencies": {
-		"@babel/core": "7.24.0",
-		"@babel/plugin-transform-typescript": "7.24.5",
-		"@babel/preset-env": "7.24.0",
-		"@babel/preset-react": "7.24.1",
-		"@babel/preset-typescript": "7.24.1",
 		"@changesets/cli": "2.27.1",
-		"@emotion/babel-plugin": "11.11.0",
 		"@guardian/eslint-config": "workspace:*",
 		"@guardian/eslint-config-typescript": "workspace:*",
 		"@guardian/prettier": "workspace:*",
 		"@manypkg/get-packages": "2.2.0",
-		"@storybook/addon-a11y": "8.0.5",
-		"@storybook/addon-docs": "8.0.5",
-		"@storybook/addon-essentials": "8.0.5",
-		"@storybook/addon-links": "8.0.5",
-		"@storybook/addon-viewport": "8.0.5",
-		"@storybook/manager-api": "^8.0.5",
-		"@storybook/react": "8.0.5",
-		"@storybook/react-webpack5": "8.0.5",
-		"@storybook/theming": "8.0.5",
 		"@swc-node/register": "1.9.0",
 		"@swc/core": "1.4.0",
 		"@swc/helpers": "^0.5.2",
-		"@types/babel__core": "7.20.5",
 		"@types/jest": "29.5.8",
 		"@types/node": "20.12.11",
 		"@typescript-eslint/eslint-plugin": "6.21.0",
 		"@typescript-eslint/parser": "6.21.0",
-		"babel-loader": "9.1.3",
 		"colorette": "2.0.20",
 		"eslint": "8.56.0",
 		"eslint-plugin-eslint-comments": "3.2.0",
@@ -49,11 +32,7 @@
 		"jest-environment-jsdom": "29.7.0",
 		"lint-staged": "15.2.0",
 		"prettier": "3.2.2",
-		"react": "18.2.0",
-		"react-dom": "18.2.0",
-		"require-from-string": "2.0.2",
 		"sort-package-json": "2.10.0",
-		"storybook": "8.0.5",
 		"ts-jest": "29.1.1",
 		"typescript": "5.3.3",
 		"update-section": "0.3.3",
